requirements.txt
asgiref==3.9.1
dj-database-url==2.1.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==25.0
psycopg2-binary
PyJWT==2.10.1
python-barcode==0.15.1
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.1
whitenoise==6.6.0